shared/requirements-core.txt
# Core dependencies shared by all deployments
pydantic-ai-slim[openai,anthropic,google,groq,huggingface]>=1.9.0
duckdb>=1.4.1
plotly>=6.3.1
tomli-w>=1.2.0
